pyproject.toml
[project]
dependencies = [
    "loguru",
    "phonenumberslite",
    "pipecat-ai-flows>=0.0.24",
    "pipecat-ai[azure,daily,rnnoise,runner,silero,webrtc]>=0.0.106",
    "pipecatcloud",
    "python-dotenv",
    "pyyaml",
    "rapidfuzz",
]
description = "Virginia Legal Aid Society (VLAS) Telephone Intake Bot"
name = "intake-bot"
readme = "README.md"
requires-python = ">=3.12"
version = "0.3.9"

[dependency-groups]
dev = [
    "fastapi",
    "pipecat-ai-cli",
    "pipecat-ai-tail",
    "pipecat-ai-whisker",
    "pipecat-ai[websocket]",
    "pytest",
    "pytest-asyncio",
    "uvicorn",
]
test = ["pytest", "pytest-asyncio"]

[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
intake_bot = ["data/*.yml"]

[tool.pytest.ini_options]
filterwarnings = ["ignore::DeprecationWarning:pipecat"]
minversion = "8.0"
pythonpath = ["src"]
required_plugins = ["pytest-asyncio"]
testpaths = ["tests"]
